requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
satpy>=0.39.0
opencv-python>=4.5.0
dill>=0.3.4
pyyaml>=6.0.0
matplotlib>=3.4.0
netCDF4>=1.5.8
h5py>=3.6.0
pyresample>=1.22.0
-e .
